requirements.txt
biopython
transformers
torch
faiss-cpu
streamlit
